♻️ refactor:  プレゼンテーション層のフォルダ構成変更
Issue #145
