+++ lib/presentation/common/pages/error_page.dart
@@ -4,8 +4,8 @@
 
 import 'package:flutter/material.dart';
 
-import '../../localizations/strings.g.dart';
-import 'components/error_view.dart';
+import '../../../localizations/strings.g.dart';
+import '../components/error_view.dart';
 
 /// エラー画面
 class ErrorPage extends StatelessWidget {
