+++ lib/presentation/repo/components/detail_view.dart
@@ -13,8 +13,8 @@ import '../../common/components/cached_circle_avatar.dart';
 import '../../common/components/error_view.dart';
 import '../../common/components/hyperlink_text.dart';
 import '../../common/components/icon_label.dart';
-import '../avatar_preview_page.dart';
-import '../view_page.dart';
+import '../pages/avatar_preview_page.dart';
+import '../pages/view_page.dart';
 import 'language_label.dart';
 import 'readme_markdown.dart';
 import 'selected_repo.dart';
