+++ lib/presentation/repo/components/search_text_button.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
-import '../search_page.dart';
+import '../pages/search_page.dart';
 import 'search_repos_query.dart';
 import 'search_text_field.dart';
