+++ lib/presentation/repo/pages/avatar_preview_page.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../utils/logger.dart';
-import 'components/avatar_preview_view.dart';
+import '../components/avatar_preview_view.dart';
 
 /// アバターのプレビュー画面
 class RepoAvatarPreviewPage extends StatelessWidget {
