+++ test/config/app_test.dart
@@ -5,7 +5,7 @@
 import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/presentation/repo/index_page.dart';
+import 'package:github_search/presentation/repo/pages/index_page.dart';
 
 import '../test_utils/hive.dart';
 import '../test_utils/mocks.dart';
