+++ test/presentation/repo/components/search_text_button_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/presentation/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/repo/components/search_text_button.dart';
-import 'package:github_search/presentation/repo/search_page.dart';
+import 'package:github_search/presentation/repo/pages/search_page.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
