+++ test/presentation/repo/components/selected_repo_test.dart
@@ -9,7 +9,7 @@ import 'package:github_search/config/router.dart';
 import 'package:github_search/domain/entities/owner.dart';
 import 'package:github_search/domain/entities/repo.dart';
 import 'package:github_search/presentation/repo/components/selected_repo.dart';
-import 'package:github_search/presentation/repo/view_page.dart';
+import 'package:github_search/presentation/repo/pages/view_page.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../test_utils/locale.dart';
