+++ test/presentation/repo/search_page_test.dart
@@ -9,7 +9,7 @@ import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/repo/components/search_text_field.dart';
 import 'package:github_search/presentation/repo/components/sort_button.dart';
 import 'package:github_search/presentation/repo/components/sort_selector_bottom_sheet.dart';
-import 'package:github_search/presentation/repo/search_page.dart';
+import 'package:github_search/presentation/repo/pages/search_page.dart';
 
 import '../../test_utils/hive.dart';
 import '../../test_utils/locale.dart';
